requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
django==5.0.4
django-browser-reload==1.12.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-js-asset==2.2.0
django-tailwind==3.8.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
jinja2==3.1.3
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
pillow==10.3.0
pygments==2.17.2
pyjwt==2.8.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pyyaml==6.0.1
requests==2.31.0
rich==13.7.1
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.1
gunicorn
